Added Map generation and introduced functionality for sprites.
In detail, I've modified the level so it's bigger, as well as adding good level generation that is (for the most part) random! Some other modifications have been made such as adding the functionality of sprites (this does not fully work yet), as well as cleaning up the GameScreen code.